+++ test/fixtures/inputForBlockModel.ts
@@ -3,6 +3,7 @@ import { IBlock } from '../../src/models/Block';
 
 export const inputForBlockModel: CreateQuery<IBlock> = {
   _id: 'block::1',
+  chainAddress: '0x1',
   anchor: '1024',
   height: 1,
   minter: '0xA405324F4b6EB7Bc76f1964489b3769cfc71445F',