+++ CHANGELOG.md
@@ -5,7 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Added
+- support for chain contract update
+- support for block reverts
+  
+### Changed
+- sync data when root is available instead of waiting for next block
+- update toolbox to `^0.9.0`
 
 ## [0.2.6] - 2021-03-17
 ### Fixed